More frequency_voltage_table tweaks, still doesn't work as expected.

+++ kernel/samsung/2.6.35/arch/arm/mach-s5pv210/cpu-freq.c
@@ -464,7 +464,7 @@ static int s5pv210_cpufreq_target(struct cpufreq_policy *policy,
 //Subtract the voltage in the undervolt table before supplying it to the cpu
 //Got to multiply by 1000 to account for the conversion between SGS and NS
 	arm_volt = (dvs_conf[index].arm_volt - (exp_UV_mV[index]*1000));
-	freq_uv_table[index][2] = arm_volt / 1000;
+	freq_uv_table[index][2] =(int) arm_volt / 1000;
 	//arm_volt = dvs_conf[index].arm_volt;
 	int_volt = dvs_conf[index].int_volt;
 
@@ -654,7 +654,7 @@ static int s5pv210_cpufreq_target(struct cpufreq_policy *policy,
 			"cpufreq: Performance changed[L%d]\n", index);
 //more uv
 	previous_arm_volt = (dvs_conf[index].arm_volt - (exp_UV_mV[index] * 1000));
-	freq_uv_table[index][2] = previous_arm_volt / 1000;
+	freq_uv_table[index][2] = (int) previous_arm_volt / 1000;
 
 	if (first_run)
 		first_run = false;
@@ -699,7 +699,7 @@ static int s5pv210_cpufreq_resume(struct cpufreq_policy *policy)
 			sizeof(struct s3c_freq));
 //even more uv
 	previous_arm_volt = (dvs_conf[level].arm_volt - (exp_UV_mV[level]*1000));
-	freq_uv_table[level][2] = previous_arm_volt / 1000;
+	freq_uv_table[level][2] = (int) previous_arm_volt / 1000;
 
 	return ret;
 }
@@ -769,7 +769,7 @@ static int __init s5pv210_cpufreq_driver_init(struct cpufreq_policy *policy)
 			sizeof(struct s3c_freq));
 //is dat some more uv?
 	previous_arm_volt = (dvs_conf[level].arm_volt - (exp_UV_mV[level]*1000));
-	freq_uv_table[level][2] = previous_arm_volt / 1000;
+	freq_uv_table[level][2] = (int) previous_arm_volt / 1000;
 
 	cpufreq_frequency_table_cpuinfo(policy, freq_table);
 //Set initial max speed to 1ghz for people who don't want to overclock
